refactor: create subfolders in /components/common folder
