+++ components/ui/home/detail/molecules/NearByPlaceCard.tsx
@@ -1,7 +1,7 @@
 import { View, Text, Image } from "react-native";
 import type { NearByPlaceData } from "@/types/api/nearby";
-import VerticalSeparator from "@/components/common/Separator";
-import StarRating from "@/components/common/StarRating";
+import VerticalSeparator from "@/components/common/display/Separator";
+import StarRating from "@/components/common/display/StarRating";
 import PlusIcon from "@/assets/icons/common/plus.svg";
 
 interface Props {
