+++ components/ui/home/detail/molecules/ReviewCard.tsx
@@ -1,9 +1,9 @@
 import { View, Text } from "react-native";
 import type { ReviewData } from "@/types/api/review";
-import StarRating from "@/components/common/StarRating";
-import LikeCount from "@/components/common/LikeCount";
-import CommentCount from "@/components/common/CommentCount";
-import UserProfile from "@/components/common/UserProfile";
+import StarRating from "@/components/common/display/StarRating";
+import LikeCount from "@/components/common/display/LikeCount";
+import CommentCount from "@/components/common/display/CommentCount";
+import UserProfile from "@/components/common/display/UserProfile";
 
 interface Props {
   review: ReviewData;
