+++ components/ui/home/detail/molecules/shared/Title.tsx
@@ -1,5 +1,5 @@
 import { View, Text } from "react-native";
-import LinkButton from "@/components/common/LinkButton";
+import LinkButton from "@/components/common/input/LinkButton";
 
 interface Props {
   title: string;
