+++ components/ui/home/index/atoms/ReviewCard.tsx
@@ -1,9 +1,9 @@
 import { Text, View, Image } from "react-native";
 import { Link } from "expo-router";
 import type { ReviewData } from "@/types/api/review";
-import CommentCount from "@/components/common/CommentCount";
-import StarRating from "@/components/common/StarRating";
-import UserProfile from "@/components/common/UserProfile";
+import CommentCount from "@/components/common/display/CommentCount";
+import StarRating from "@/components/common/display/StarRating";
+import UserProfile from "@/components/common/display/UserProfile";
 
 interface Props {
   data: ReviewData;
