+++ components/ui/travelReview/molecules/Comment.tsx
@@ -1,7 +1,7 @@
 import { Text, View } from "react-native";
 import type { CommentResDto, Reply } from "@/types/api/review";
-import UserProfile from "@/components/common/UserProfile";
-import ReplyCount from "@/components/common/ReplyCount";
+import UserProfile from "@/components/common/display/UserProfile";
+import ReplyCount from "@/components/common/display/ReplyCount";
 import CommentBar from "@/components/ui/travelReview/molecules/CommentBar";
 
 interface Props {
